% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday Function}
\usage{
birthday(x)
}
\arguments{
\item{x}{# The number of people}
}
\value{
# Returns the probability of 2 people in the sample having the same birthday
}
\description{
Birthday Function
}
\examples{
birthday(20:24)
}
