% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
## `fire`
A data frame with 15 rows and 2 columns
\describe{
  \item{Distance}{Distance the fire spread}
  \item{Damage}{Cost of the damage inflicted}
}
}
\source{
Local course resource: FIREDAM.csv
}
\usage{
fire
}
\description{
A data set of fire incidents and damages
}
\keyword{datasets}
