% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{mybin}
\alias{mybin}
\title{Title}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{#Number of trials per experiment.}

\item{n}{#Probability of success for each trial.}

\item{p}{#Number of simulation iterations(experiments).}
}
\value{
# A barplot showing the number of successes vs the frequency.
}
\description{
Title
}
\examples{
mybin(100, 10, 0.8)
}
