% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overbooking.R
\name{ntickets}
\alias{ntickets}
\title{Discrete and normal approximations to the
overbooking problem}
\usage{
ntickets(N, gamma, p, max_mult = 5, plot = TRUE)
}
\arguments{
\item{N}{number of seats}

\item{gamma}{desired maximum overbooking
probability}

\item{p}{probability of people showing up}

\item{max_mult}{maximum for search upper
bound}

\item{plot}{Should plots be shown}
}
\value{
a list with elements nd, nc, N, p,
gamma, and two  plots showing the discrete
and continuous cases
}
\description{
Discrete and normal approximations to the
overbooking problem
}
\examples{
ntickets(400, 0.02, 0.95, 5, TRUE)
}
